Fixed parameter naming problem in nmixture registration

+++ R/dNmixture.R
@@ -144,8 +144,8 @@ dNmixture_v <- nimbleFunction(
     run = function(x = double(1),
                    lambda = double(),
                    prob = double(1),
-                   Nmin = integer(0, default = -1),
-                   Nmax = integer(0, default = -1),
+                   Nmin = double(0, default = -1),
+                   Nmax = double(0, default = -1),
                    len = double(),
                    log = integer(0, default = 0)) {
     if (length(x) != len) stop("in dNmixture_v, len must equal length(x).")
@@ -195,8 +195,8 @@ dNmixture_s <- nimbleFunction(
     run = function(x = double(1),
                    lambda = double(),
                    prob = double(),
-                   Nmin = integer(0, default = -1),
-                   Nmax = integer(0, default = -1),
+                   Nmin = double(0, default = -1),
+                   Nmax = double(0, default = -1),
                    len = double(),
                    log = integer(0, default = 0)) {
     if (length(x) != len) stop("in dNmixture_v, len must equal length(x).")
@@ -245,8 +245,8 @@ rNmixture_v <- nimbleFunction(
   run = function(n = double(),
                  lambda = double(),
                  prob = double(1),
-                 Nmin = integer(0, default = -1),
-                 Nmax = integer(0, default = -1),
+                 Nmin = double(0, default = -1),
+                 Nmax = double(0, default = -1),
                  len = double()) {
     if (n != 1) stop("rNmixture_v only works for n = 1")
     if (length(prob) != len) stop("In rNmixture_v, len must equal length(prob).")
@@ -267,8 +267,8 @@ rNmixture_s <- nimbleFunction(
   run = function(n = double(),
                  lambda = double(),
                  prob = double(),
-                 Nmin = integer(0, default = -1),
-                 Nmax = integer(0, default = -1),
+                 Nmin = double(0, default = -1),
+                 Nmax = double(0, default = -1),
                  len = double()) {
     if (n != 1) stop("rNmixture_v only works for n = 1")
     trueN <- rpois(1, lambda)
